Extract config-reload lifecycle from orchestrator into orchestration/config_lifecycle
ConfigReloadLifecycle now owns the debounce/drain state machine, the reload
loop, and the load->diff->plan sequencing. The orchestrator wires the
collaborator and keeps ownership of applying a plan (restarting bots,
accounts, MCP, rooms) via _apply_config_update_plan. The pure helpers
configured_entity_names and plugin_change_paths move to
orchestration/config_updates.py.
